Renaming files to comport with suggested convention.
