+++ src/commit-order.cpp
@@ -1,4 +1,4 @@
-#include "CommitOrder.h"
+#include "commit-order.h"
 
 #include <algorithm>
